Add SPlitPayment as separate package
