+++ src/Cielo/API30/Ecommerce/SplitPayment/SplitPayment.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Cielo\API30\Ecommerce;
+namespace Cielo\API30\Ecommerce\SplitPayment;
 
 /**
  * Class SplitPayment
