+++ src/Cielo/API30/Ecommerce/SplitPayment/SplitPaymentFares.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Cielo\API30\Ecommerce;
+namespace Cielo\API30\Ecommerce\SplitPayment;
 
 /**
  * Class SplitPaymentFares
